pyproject.toml
[project]
name = "bullpen"
version = "0.1.0"
description = "The Bullpen — AI agent platform for Texas CRE intelligence"
requires-python = ">=3.11"
dependencies = [
    "prefect>=3.0",
    "supabase>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "apscheduler>=3.10,<4.0",
    "python-jobspy>=1.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
